pyproject.toml
[tool.poetry]
name = "task-list-manager"
version = "0.1.0"
description = ""
authors = ["kashmir <kuznetcov.alekcey@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.17"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
